requirements.txt
opencv-python>=4.8.0
pillow>=10.0.0
webp>=0.1.5
